refactor parser unit test, moved fixtures out to seperate file
