+++ src/test/fixtures/es5ReactComponentFixtures.js
@@ -1,0 +1,17 @@
+module.exports = {
+  singleMainApp: `var Main = React.createClass({ })`,
+  nestedComponents: `
+  var Main = React.createClass({
+        render: function() {
+          return <div>
+            <SearchBar />
+            <div>Testing</div>
+            <SearchResults>
+              <Result />
+              <Result />
+            </SearchResults>
+          </div>
+        }
+      });
+  `,
+}
